move shared RoundedLineBorder to plugin-lib
